+++ internal/xsql/conn.go
@@ -15,7 +15,6 @@ import (
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xcontext"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql/badconn"
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql/isolation"
 	"github.com/ydb-platform/ydb-go-sdk/v3/retry"
 	"github.com/ydb-platform/ydb-go-sdk/v3/scheme"
 	"github.com/ydb-platform/ydb-go-sdk/v3/table"
@@ -25,6 +24,14 @@ import (
 )
 
 type connOption func(*conn)
+
+func withFakeTxModes(modes ...QueryMode) connOption {
+	return func(c *conn) {
+		for _, m := range modes {
+			c.beginTxFuncs[m] = c.beginTxFake
+		}
+	}
+}
 
 func withDataOpts(dataOpts ...options.ExecuteDataQueryOption) connOption {
 	return func(c *conn) {
@@ -56,10 +63,14 @@ func withTrace(t trace.DatabaseSQL) connOption {
 	}
 }
 
+type beginTxFunc func(ctx context.Context, txOptions driver.TxOptions) (currentTx, error)
+
 type conn struct {
 	connector *Connector
 	trace     trace.DatabaseSQL
 	session   table.ClosableSession // Immutable and r/o usage.
+
+	beginTxFuncs map[QueryMode]beginTxFunc
 
 	closed           uint32
 	lastUsage        int64
@@ -115,6 +126,9 @@ func newConn(c *Connector, s table.ClosableSession, opts ...connOption) *conn {
 	cc := &conn{
 		connector: c,
 		session:   s,
+	}
+	cc.beginTxFuncs = map[QueryMode]beginTxFunc{
+		DataQueryMode: cc.beginTx,
 	}
 	for _, o := range opts {
 		if o != nil {
@@ -391,39 +405,32 @@ func (c *conn) normalize(q string, args ...driver.NamedValue) (query string, _ *
 	}()...)
 }
 
-func (c *conn) BeginTx(ctx context.Context, txOptions driver.TxOptions) (_ driver.Tx, err error) {
+func (c *conn) BeginTx(ctx context.Context, txOptions driver.TxOptions) (tx driver.Tx, err error) {
 	var transaction table.Transaction
 	onDone := trace.DatabaseSQLOnConnBegin(c.trace, &ctx)
 	defer func() {
 		onDone(transaction, err)
 	}()
-	m := queryModeFromContext(ctx, c.defaultQueryMode)
-	if m != DataQueryMode {
-		return nil, badconn.Map(xerrors.WithStackTrace(fmt.Errorf("wrong query mode: %s", m.String())))
-	}
-	if !c.isReady() {
-		return nil, badconn.Map(xerrors.WithStackTrace(errNotReadyConn))
-	}
+
 	if c.currentTx != nil {
 		return nil, xerrors.WithStackTrace(
 			fmt.Errorf("conn already have an opened currentTx: %s", c.currentTx.ID()),
 		)
 	}
-	var txc table.TxOption
-	txc, err = isolation.ToYDB(txOptions)
+
+	m := queryModeFromContext(ctx, c.defaultQueryMode)
+
+	beginTx, isKnown := c.beginTxFuncs[m]
+	if !isKnown {
+		return nil, badconn.Map(xerrors.WithStackTrace(fmt.Errorf("wrong query mode: %s", m.String())))
+	}
+
+	tx, err = beginTx(ctx, txOptions)
 	if err != nil {
 		return nil, xerrors.WithStackTrace(err)
 	}
-	transaction, err = c.session.BeginTransaction(ctx, table.TxSettings(txc))
-	if err != nil {
-		return nil, badconn.Map(xerrors.WithStackTrace(err))
-	}
-	c.currentTx = &tx{
-		conn: c,
-		ctx:  ctx,
-		tx:   transaction,
-	}
-	return c.currentTx, nil
+
+	return tx, nil
 }
 
 func (c *conn) Version(context.Context) (_ string, err error) {