+++ internal/xsql/tx_fake.go
@@ -1,0 +1,91 @@
+package xsql
+
+import (
+	"context"
+	"database/sql/driver"
+
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xcontext"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xsql/badconn"
+	"github.com/ydb-platform/ydb-go-sdk/v3/table"
+	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
+)
+
+type txFake struct {
+	conn *conn
+	ctx  context.Context
+}
+
+var (
+	_ driver.Tx                   = &txFake{}
+	_ driver.ExecerContext        = &txFake{}
+	_ driver.QueryerContext       = &txFake{}
+	_ table.TransactionIdentifier = &txFake{}
+)
+
+func (c *conn) beginTxFake(ctx context.Context, txOptions driver.TxOptions) (currentTx, error) {
+	return &txFake{
+		conn: c,
+		ctx:  ctx,
+	}, nil
+}
+
+func (tx *txFake) ID() string {
+	return "FAKE"
+}
+
+func (tx *txFake) Commit() (err error) {
+	onDone := trace.DatabaseSQLOnTxCommit(tx.conn.trace, &tx.ctx, tx)
+	defer func() {
+		onDone(err)
+	}()
+	defer func() {
+		tx.conn.currentTx = nil
+	}()
+	if !tx.conn.isReady() {
+		return badconn.Map(xerrors.WithStackTrace(errNotReadyConn))
+	}
+	return nil
+}
+
+func (tx *txFake) Rollback() (err error) {
+	onDone := trace.DatabaseSQLOnTxRollback(tx.conn.trace, &tx.ctx, tx)
+	defer func() {
+		onDone(err)
+	}()
+	defer func() {
+		tx.conn.currentTx = nil
+	}()
+	if !tx.conn.isReady() {
+		return badconn.Map(xerrors.WithStackTrace(errNotReadyConn))
+	}
+	return err
+}
+
+func (tx *txFake) QueryContext(ctx context.Context, query string, args []driver.NamedValue) (
+	rows driver.Rows, err error,
+) {
+	onDone := trace.DatabaseSQLOnTxQuery(tx.conn.trace, &ctx, tx.ctx, tx, query, xcontext.IsIdempotent(ctx))
+	defer func() {
+		onDone(err)
+	}()
+	rows, err = tx.conn.QueryContext(ctx, query, args)
+	if err != nil {
+		return nil, xerrors.WithStackTrace(err)
+	}
+	return rows, nil
+}
+
+func (tx *txFake) ExecContext(ctx context.Context, query string, args []driver.NamedValue) (
+	result driver.Result, err error,
+) {
+	onDone := trace.DatabaseSQLOnTxExec(tx.conn.trace, &ctx, tx.ctx, tx, query, xcontext.IsIdempotent(ctx))
+	defer func() {
+		onDone(err)
+	}()
+	result, err = tx.conn.ExecContext(ctx, query, args)
+	if err != nil {
+		return nil, xerrors.WithStackTrace(err)
+	}
+	return result, nil
+}